backend/pipelines/video-annotator-mvp/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
pydantic
python-dotenv
ultralytics
scipy
modal
httpx
